修复!：群宠下的宠物顺序，宠物进化，宠物技能
1. 宠物召唤，允许召唤多个宝贝龙/机器人；
    - 原效果：只要背包存在，就无法召唤
2. 修复任务4659，根据Check.img修改脚本收取任务道具而不是金币；
    - 原效果：消耗10000金币
    - 新效果：消耗 4000111 * 50
3. 优化群宠下宠物进化，提供选择指定的宠物进行后续操作；
    - 原效果：通过itemId找到最先满足条件的宠物，不过由于有1.的限制，这个操作没有带来问题
    - 新效果：通过对话框选择指定的宠物
4. 补充宠物进化时的特效；
5. 修复群宠下，宠物进化后的顺序：
    - 原效果：宠物顺序将由于**召回进化前-重排-召唤进化后**而导致顺序发生改变
    - 新效果：宠物顺序保持不变
6. 修复上线时的宠物顺序
    - 原效果：宠物顺序会按照其在背包中的顺序被召唤
    - 新效果：宠物顺序保持下线前的顺序
7. 宠物进化概率数据改用wz中的数据
    - 新效果：导致部分稀有造型概率大幅度下降（大约从10%下降到1%）
8. 宠物进化过期时间继承
    - 原效果：宠物进化后过期时间刷新（90天）
    - 新效果：宠物进化继承原宠物的过期时间
9. 修复不同职业群学习的群宠技能不同
10. 修复任务4660, 4661 学习的宠物技能

!! 由于数据库先前并没有记录宠物顺序，将导致所有宠物被召回。

+++ gms-server/src/main/java/org/gms/scripting/npc/NPCScriptManager.java
@@ -169,7 +169,7 @@ public class NPCScriptManager extends AbstractScriptManager {
         }
     }
 
-    public void action(Client c, byte mode, byte type, int selection) {
+    public void action(Client c, byte mode, byte type, long selection) {
         Invocable iv = scripts.get(c);
         if (iv != null) {
             try {
@@ -187,7 +187,7 @@ public class NPCScriptManager extends AbstractScriptManager {
         }
     }
 
-    public void nextLevel(Client c, byte mode, byte type, int selection) {
+    public void nextLevel(Client c, byte mode, byte type, long selection) {
         Invocable iv = scripts.get(c);
         if (iv != null) {
             try {
